requirements.txt
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
pillow==10.1.0
cmake==3.27.7
dlib==19.24.2
